requirements.txt
psycopg2-binary==2.9.5
bs4==0.0.1
requests==2.28.1
numpy==1.23.2
urllib3==1.26.12
lxml==4.9.1
PyNaCl==1.5.0
furl==2.1.3